# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Inventory Asset Core API
  version: 0.20.0
servers:
  - url: /api/v1
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
    bearerToken:
      type: http
      scheme: bearer
  schemas:
    Observation:
      type: object
      required: [source, external_id, data]
      properties:
        source: {type: string, maxLength: 64}
        external_id: {type: string, maxLength: 512}
        data: {type: object, additionalProperties: true}
        normalized: {type: object, additionalProperties: true}
paths:
  /assets:
    get:
      security: [{sessionCookie: []}]
      parameters:
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 2000}}
        - {name: offset, in: query, schema: {type: integer, minimum: 0}}
        - {name: active, in: query, schema: {type: string}}
        - {name: status, in: query, schema: {type: string}}
        - {name: source, in: query, schema: {type: string}}
      responses:
        "200": {description: Asset list}
  /assets/{asset_uuid}:
    get:
      security: [{sessionCookie: []}]
      parameters:
        - {name: asset_uuid, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        "200": {description: Asset and source states}
        "404": {description: Not found}
  /assets/{asset_uuid}/changes:
    get:
      security: [{sessionCookie: []}]
      parameters:
        - {name: asset_uuid, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        "200": {description: Temporal changes}
  /assets/{asset_uuid}/snapshots:
    get:
      security: [{sessionCookie: []}]
      parameters:
        - {name: asset_uuid, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        "200": {description: Snapshots}
  /assets/observations:
    post:
      security: [{bearerToken: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/Observation"}
      responses:
        "201": {description: Observation accepted}
        "400": {description: Invalid payload}
        "401": {description: Invalid token or rate limit}
  /assets/observations/batch:
    post:
      security: [{bearerToken: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [items]
              properties:
                items:
                  type: array
                  maxItems: 500
                  items: {$ref: "#/components/schemas/Observation"}
      responses:
        "201": {description: Batch accepted}
        "207": {description: Partial success}
  /sources:
    get:
      security: [{sessionCookie: []}]
      responses:
        "200": {description: Source health/state}
  /sources/sync:
    post:
      security: [{sessionCookie: []}]
      responses:
        "202": {description: Sync started}
  /tokens:
    get:
      security: [{sessionCookie: []}]
      responses:
        "200": {description: API token metadata}
    post:
      security: [{sessionCookie: []}]
      responses:
        "201": {description: Token generated; plaintext returned once}
  /tokens/{token_id}:
    delete:
      security: [{sessionCookie: []}]
      parameters:
        - {name: token_id, in: path, required: true, schema: {type: integer}}
      responses:
        "200": {description: Token revoked}
  /outbox:
    get:
      security: [{sessionCookie: []}]
      responses:
        "200": {description: Event backlog}
  /dead-letters:
    get:
      security: [{sessionCookie: []}]
      responses:
        "200": {description: Dead-letter events}
  /dead-letters/{dead_id}/retry:
    post:
      security: [{sessionCookie: []}]
      parameters:
        - {name: dead_id, in: path, required: true, schema: {type: integer}}
      responses:
        "200": {description: Event requeued}